Split Pulse dashboard sidebar
